# src/Balsam/api/openapi.yaml
openapi: 3.0.2
info:
  description: This a service contract for the OicdProvider in Balsam.
  title: S3Provider
  version: "2.0"
servers:
- url: http://s3-provider.balsam-system.svc.cluster.local/api/v1
tags:
- description: Buckets methods
  name: Bucket
paths:
  /buckets:
    post:
      description: Creates a new Bucket
      operationId: CreateBucket
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBucketRequest'
        description: Definition of a new role
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BucketCreatedResponse'
          description: Success
        "400":
          content:
            application/problem+json:
              example:
                title: Bad Request
                status: 400
                detail: Generic error
              schema:
                $ref: '#/components/schemas/Problem'
          description: Error respsone for 400
      tags:
      - Bucket
  /buckets/{bucketId}/folder:
    post:
      description: Creates a new virtual folder in the bucket.
      operationId: CreateFolder
      parameters:
      - description: the name of the bucket
        explode: false
        in: path
        name: bucketId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
        description: Definition of a new role
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderCreatedResponse'
          description: Success
        "400":
          content:
            application/problem+json:
              example:
                title: Bad Request
                status: 400
                detail: Generic error
              schema:
                $ref: '#/components/schemas/Problem'
          description: Error respsone for 400
      tags:
      - Bucket
  /buckets/{bucketId}/acceskey:
    post:
      description: Creates a new acces key for the bucket.
      operationId: CreateAccessKey
      parameters:
      - description: the name of the bucket
        explode: false
        in: path
        name: bucketId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKeyCreatedResponse'
          description: Success
        "400":
          content:
            application/problem+json:
              example:
                title: Bad Request
                status: 400
                detail: Generic error
              schema:
                $ref: '#/components/schemas/Problem'
          description: Error respsone for 400
      tags:
      - Bucket
components:
  responses:
    ErrorResponse400:
      content:
        application/problem+json:
          example:
            title: Bad Request
            status: 400
            detail: Generic error
          schema:
            $ref: '#/components/schemas/Problem'
      description: Error respsone for 400
  schemas:
    CreateBucketRequest:
      description: Bucket name
      example:
        policyName: policyName
        name: name
      properties:
        name:
          description: The name of the bucket
          type: string
        policyName:
          description: name of the s3 policy name should match the group name in Oidc
            provider
          type: string
      required:
      - name
      - policyName
      type: object
    CreateFolderRequest:
      description: Folder name
      example:
        name: name
      properties:
        name:
          description: The name of the folder
          type: string
      required:
      - name
      type: object
    BucketCreatedResponse:
      description: Bucket description
      example:
        requestedName: requestedName
        name: name
      properties:
        requestedName:
          description: The requested name of the bucket
          type: string
        name:
          description: The name of the bucket
          type: string
      required:
      - name
      - requestedName
      type: object
    FolderCreatedResponse:
      description: Folder created description
      example:
        requestedName: requestedName
        name: name
      properties:
        requestedName:
          description: The requested name of the virtual folder
          type: string
        name:
          description: The name of the virtual folder
          type: string
      required:
      - name
      - requestedName
      type: object
    AccessKeyCreatedResponse:
      description: Access key created description
      example:
        secretKey: secretKey
        accessKey: accessKey
      properties:
        accessKey:
          description: The access key
          type: string
        secretKey:
          description: The secret key
          type: string
      required:
      - accessKey
      - secretKey
      type: object
    Problem:
      properties:
        type:
          default: about:blank
          description: |
            An absolute URI that identifies the problem type.  When dereferenced,
            it SHOULD provide human-readable documentation for the problem type
            (e.g., using HTML).
          example: https://zalando.github.io/problem/constraint-violation
          format: uri
          type: string
        title:
          description: |
            A short, summary of the problem type. Written in english and readable
            for engineers (usually not suited for non technical stakeholders and
            not localized); example: Service Unavailable
          type: string
        status:
          description: |
            The HTTP status code generated by the origin server for this occurrence
            of the problem.
          example: 503
          exclusiveMaximum: true
          format: int32
          maximum: 600
          minimum: 100
          type: integer
        detail:
          description: |
            A human readable explanation specific to this occurrence of the
            problem.
          example: Connection to database timed out
          type: string
        instance:
          description: "An absolute URI that identifies the specific occurrence of\
            \ the problem.\nIt may or may not yield further information if dereferenced.\n\
            \      \n"
          format: uri
          type: string
      type: object